requirements.txt
Click==7.0
Flask==1.0.2
Flask-Assets==0.12
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
libsass==0.15.1
MarkupSafe==1.0
pyScss==1.3.5
python-dotenv==0.9.1
six==1.11.0
webassets==0.12.1
Werkzeug==0.14.1
